+++ Source/Patches/PhysGrabberPatches.cs
@@ -17,24 +17,6 @@ namespace RepoXR.Patches;
 [RepoXRPatch]
 internal static class PhysGrabberPatches
 {
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static Transform GetHandTransform()
-    {
-        if (VRSession.Instance is { } session)
-            return session.Player.MainHand;
-
-        return PhysGrabber.instance.playerCamera.transform;
-    }
-
-    private static CodeMatcher ReplaceCameraWithHand(this CodeMatcher matcher)
-    {
-        var labels = matcher.Instruction.labels;
-
-        return matcher.RemoveInstructions(2).InsertAndAdvance(
-            new CodeInstruction(OpCodes.Call, ((Func<Transform>)GetHandTransform).Method).WithLabels(labels)
-        );
-    }
-
     /// <summary>
     /// Make certain phys grabber operations operate from the hand transform instead of the camera transform
     /// </summary>
@@ -44,16 +26,8 @@ internal static class PhysGrabberPatches
     {
         return new CodeMatcher(instructions)
             .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PhysGrabber), nameof(PhysGrabber.playerCamera))))
-            .Repeat(matcher => matcher.Advance(-1).ReplaceCameraWithHand())
-            .Start()
-            .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PlayerAvatar), nameof(PlayerAvatar.PlayerVisionTarget))))
-            .Advance(-2)
-            .RemoveInstructions(4)
-            .Insert(
-                new CodeInstruction(OpCodes.Call, ((Func<Transform>)GetHandTransform).Method)
-            )
+                new CodeMatch(OpCodes.Callvirt, Method(typeof(PlayerLocalCamera), nameof(PlayerLocalCamera.GetOverrideTransform))))
+            .Repeat(matcher => matcher.Set(OpCodes.Call, PlayerLocalCameraExtensions.GetHandOverrideTransformMethod))
             .InstructionEnumeration();
     }
 
@@ -67,23 +41,19 @@ internal static class PhysGrabberPatches
         return new CodeMatcher(instructions)
             .MatchForward(false, new CodeMatch(OpCodes.Ldc_R4, 0.2f))
             .SetOperandAndAdvance(0.1f)
-            .Advance(1)
             .Insert(
                 new CodeInstruction(OpCodes.Ldsfld, Field(typeof(InputManager), nameof(InputManager.instance))),
                 new CodeInstruction(OpCodes.Callvirt,
                     Method(typeof(InputManager), nameof(InputManager.KeyPullAndPush))),
-                new CodeInstruction(OpCodes.Call,
-                    Method(typeof(Vector3), "op_Multiply", [typeof(Vector3), typeof(float)]))
+                new CodeInstruction(OpCodes.Mul)
             )
             .MatchForward(false, new CodeMatch(OpCodes.Ldc_R4, 0.2f))
             .SetOperandAndAdvance(-0.1f)
-            .Advance(1)
             .Insert(
                 new CodeInstruction(OpCodes.Ldsfld, Field(typeof(InputManager), nameof(InputManager.instance))),
                 new CodeInstruction(OpCodes.Callvirt,
                     Method(typeof(InputManager), nameof(InputManager.KeyPullAndPush))),
-                new CodeInstruction(OpCodes.Call,
-                    Method(typeof(Vector3), "op_Multiply", [typeof(Vector3), typeof(float)]))
+                new CodeInstruction(OpCodes.Mul)
             )
             .InstructionEnumeration();
     }
@@ -96,15 +66,20 @@ internal static class PhysGrabberPatches
     /// This is normally done by having these be a child of the camera, however this doesn't work in VR since
     /// we use our hand to move items, not the main camera.
     /// </para>
+    ///
+    /// <para>
+    /// We do not touch the transforms if a camera override is active (walkie-talkie)
+    /// </para>
     /// </summary>
     [HarmonyPatch(typeof(PhysGrabber), nameof(PhysGrabber.Update))]
     [HarmonyPostfix]
     private static void UpdatePhysGrabPlane(PhysGrabber __instance)
     {
-        if (!__instance.isLocal || !__instance.grabbedObjectTransform)
+        if (!__instance.isLocal || !__instance.grabbedObjectTransform ||
+            __instance.playerAvatar.localCamera.GetOverrideActive())
             return;
 
-        var hand = GetHandTransform();
+        var hand = __instance.playerAvatar.localCamera.GetHandOverrideTransform();
         var distancePlane = Vector3.Distance(hand.position, __instance.physGrabPointPlane.position);
         var distancePuller = Vector3.Distance(hand.position, __instance.physGrabPointPuller.position);
 
@@ -144,47 +119,9 @@ internal static class PhysGrabberPatches
     {
         return new CodeMatcher(instructions)
             .MatchForward(false,
-                new CodeMatch(OpCodes.Call,
-                    Method(typeof(Physics), nameof(Physics.Raycast),
-                    [
-                        typeof(Vector3), typeof(Vector3), typeof(RaycastHit).MakeByRefType(), typeof(float),
-                        typeof(int),
-                        typeof(QueryTriggerInteraction)
-                    ])))
-            .Advance(-14)
-            .InsertAndAdvance(
-                new CodeInstruction(OpCodes.Call, ((Func<PhysGrabber, Vector3>)CalculateNewForward).Method),
-                new CodeInstruction(OpCodes.Stloc_1),
-                new CodeInstruction(OpCodes.Ldarg_0)
-            )
-            .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PhysGrabber), nameof(PhysGrabber.playerCamera))))
-            .Repeat(matcher => matcher.Advance(-1).ReplaceCameraWithHand())
-            .InstructionEnumeration();
-
-        static Vector3 CalculateNewForward(PhysGrabber grabber)
-        {
-            if (grabber.overrideGrabTarget)
-                return (grabber.overrideGrabTarget.transform.position - VRSession.Instance.Player.MainHand.position)
-                    .normalized;
-
-            return VRSession.Instance is not { } session ? Vector3.zero : session.Player.MainHand.forward;
-        }
-    }
-
-    /// <summary>
-    /// Make "scrolling" update the position based on the hand, instead of the camera
-    /// </summary>
-    [HarmonyPatch(typeof(PhysGrabber), nameof(PhysGrabber.OverridePullDistanceIncrement))]
-    [HarmonyTranspiler]
-    private static IEnumerable<CodeInstruction> OverridePullDistanceIncrementPatches(
-        IEnumerable<CodeInstruction> instructions)
-    {
-        return new CodeMatcher(instructions)
-            .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PhysGrabber), nameof(PhysGrabber.playerCamera))))
-            .Advance(-1)
-            .ReplaceCameraWithHand()
+                new CodeMatch(OpCodes.Callvirt,
+                    Method(typeof(PlayerLocalCamera), nameof(PlayerLocalCamera.GetOverrideTransform))))
+            .Set(OpCodes.Call, PlayerLocalCameraExtensions.GetHandOverrideTransformMethod)
             .InstructionEnumeration();
     }
 
@@ -213,20 +150,6 @@ internal static class PhysGrabberPatches
 
         x = input.x;
         y = input.y;
-    }
-
-    /// <summary>
-    /// Move the grab beam origin to the hand
-    /// </summary>
-    [HarmonyPatch(typeof(PhysGrabBeam), nameof(PhysGrabBeam.Start))]
-    [HarmonyPostfix]
-    private static void OnPhysBeamStart(PhysGrabBeam __instance)
-    {
-        if (!__instance.playerAvatar.isLocal || VRSession.Instance is not {} session)
-            return;
-
-        __instance.PhysGrabPointOrigin.SetParent(session.Player.MainHand);
-        __instance.PhysGrabPointOrigin.localPosition = Vector3.zero;
     }
 
     /// <summary>
@@ -298,34 +221,6 @@ internal static class PhysGrabberPatches
 [RepoXRPatch(RepoXRPatchTarget.Universal)]
 internal static class PhysGrabberUniversalPatches
 {
-    private static Transform GetHandTransform(PhysGrabber grabber)
-    {
-        if (grabber.playerAvatar.isLocal)
-            return VRSession.InVR ? VRSession.Instance.Player.MainHand : grabber.playerAvatar.localCamera.transform;
-
-        if (!NetworkSystem.instance)
-            return grabber.playerAvatar.localCamera.transform;
-
-        return NetworkSystem.instance.GetNetworkPlayer(grabber.playerAvatar, out var networkPlayer)
-            ? networkPlayer.PrimaryHand
-            : grabber.playerAvatar.localCamera.transform;
-    }
-
-    /// <summary>
-    /// Make certain phys grabber logic be applied based on the hand instead of the head
-    /// </summary>
-    [HarmonyPatch(typeof(PhysGrabber), nameof(PhysGrabber.PhysGrabLogic))]
-    [HarmonyTranspiler]
-    private static IEnumerable<CodeInstruction> PhysGrabLogicPatches(IEnumerable<CodeInstruction> instructions)
-    {
-        return new CodeMatcher(instructions)
-            .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PhysGrabber), nameof(PhysGrabber.playerCamera))))
-            .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Call,
-                ((Func<PhysGrabber, Transform>)GetHandTransform).Method))
-            .InstructionEnumeration();
-    }
-
     [HarmonyPatch(typeof(PhysGrabber), nameof(PhysGrabber.ObjectTurning))]
     [HarmonyTranspiler]
     private static IEnumerable<CodeInstruction> ObjectTurningPatches(IEnumerable<CodeInstruction> instructions)
@@ -333,17 +228,14 @@ internal static class PhysGrabberUniversalPatches
         return new CodeMatcher(instructions)
             // Replace camera transform with hand transform (local player)
             .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PlayerAvatar), nameof(PlayerAvatar.localCamera))))
-            .Advance(-1)
-            .RemoveInstructions(3)
-            .Insert(new CodeInstruction(OpCodes.Call, ((Func<PhysGrabber, Transform>)GetHandTransform).Method))
+                new CodeMatch(OpCodes.Callvirt,
+                    Method(typeof(PlayerLocalCamera), nameof(PlayerLocalCamera.GetOverrideTransform))))
+            .Set(OpCodes.Call, PlayerLocalCameraExtensions.GetHandOverrideTransformMethod)
             // Replace camera transform with hand transform (remote player)
             .MatchForward(false,
-                new CodeMatch(OpCodes.Ldfld, Field(typeof(PlayerAvatar), nameof(PlayerAvatar.localCamera))))
-            .Advance(-1)
-            .RemoveInstructions(3)
-            .Insert(new CodeInstruction(OpCodes.Call,
-                ((Func<PhysGrabber, Transform>)GetHandTransform).Method))
+                new CodeMatch(OpCodes.Callvirt,
+                    Method(typeof(PlayerLocalCamera), nameof(PlayerLocalCamera.GetOverrideTransform))))
+            .Set(OpCodes.Call, PlayerLocalCameraExtensions.GetHandOverrideTransformMethod)
             .InstructionEnumeration();
     }