+++ Source/Patches/Player/MapToolPatches.cs
@@ -31,9 +31,11 @@ internal static class MapToolPatches
     [HarmonyTranspiler]
     private static IEnumerable<CodeInstruction> MapToolDisableInput(IEnumerable<CodeInstruction> instructions)
     {
+        // TODO: This will __definitely__ break in non-tester builds (I'm targeting a useless intermediary variable)
         return new CodeMatcher(instructions)
             .Start()
-            .RemoveInstructions(Debug.isDebugBuild ? 149 : 89)
+            .Advance(11)
+            .Set(OpCodes.Ldc_I4_0, null)
             .InstructionEnumeration();
     }
 
@@ -147,30 +149,51 @@ internal static class UniversalMapToolPatches
     [HarmonyTranspiler]
     private static IEnumerable<CodeInstruction> MapToolTransformPatch(IEnumerable<CodeInstruction> instructions)
     {
-        var shouldMutate = new CodeInstruction(OpCodes.Call, ((Func<PhotonView, bool>)ShouldMutateTransforms).Method);
+        var shouldMutateInstr =
+            new CodeInstruction(OpCodes.Call, ((Func<PhotonView, bool>)ShouldMutateTransforms).Method);
+        var isLocalOrVRPlayerInstr =
+            new CodeInstruction(OpCodes.Call, ((Func<bool, PhotonView, bool>)IsLocalOrVRPlayer).Method);
+        var isLocalPlayerMatch = new CodeMatch(OpCodes.Ldloc_0);
 
         return Debug.isDebugBuild
+            // TODO: This branch is now very confusing in this version, requires testing!
             ? new CodeMatcher(instructions)
+                // Patch hardcoded 90-degree angle to account for VR usage
                 .MatchForward(false, new CodeMatch(OpCodes.Ldc_R4, 90f))
                 .SetAndAdvance(OpCodes.Ldarg_0, null)
                 .Insert(new CodeInstruction(OpCodes.Call, ((Func<MapToolController, float>)GetHoldAngle).Method))
-                .MatchForward(false,
-                    new CodeMatch(OpCodes.Callvirt, PropertyGetter(typeof(PhotonView), nameof(PhotonView.IsMine))))
-                .SetInstructionAndAdvance(shouldMutate)
-                .MatchForward(false,
-                    new CodeMatch(OpCodes.Callvirt, PropertyGetter(typeof(PhotonView), nameof(PhotonView.IsMine))))
-                .SetInstructionAndAdvance(shouldMutate)
+                // Do not reset _hideAmount if target player is in VR
+                .MatchForward(false, isLocalPlayerMatch)
+                .SetAndAdvance(OpCodes.Ldloc_0, null)
+                .InsertAndAdvance(
+                    new CodeInstruction(OpCodes.Ldarg_0),
+                    new CodeInstruction(OpCodes.Ldfld, Field(typeof(MapToolController), nameof(MapToolController.photonView))),
+                    isLocalOrVRPlayerInstr
+                )
+                // Do not reset _speedMult if target player is in VR
+                .MatchForward(false, isLocalPlayerMatch)
+                .SetAndAdvance(OpCodes.Ldloc_0, null)
+                .InsertAndAdvance(
+                    new CodeInstruction(OpCodes.Ldarg_0),
+                    new CodeInstruction(OpCodes.Ldfld, Field(typeof(MapToolController), nameof(MapToolController.photonView))),
+                    isLocalOrVRPlayerInstr
+                )
+                // Do not apply FollowTransformClient transforms if target player is in VR
                 .MatchForward(false,
                     new CodeMatch(OpCodes.Ldfld,
                         Field(typeof(MapToolController), nameof(MapToolController.FollowTransformClient))))
-                .Advance(-18) // PhotonView::get_IsMine
-                .SetInstructionAndAdvance(shouldMutate)
-                .SetOpcodeAndAdvance(OpCodes.Brtrue_S)
-                // Nop-ify the else branch since it resets transforms
-                // TODO: Might need to be added to the non-tester branch as well
+                .MatchBack(false, isLocalPlayerMatch)
+                .SetAndAdvance(OpCodes.Ldloc_0, null)
+                .InsertAndAdvance(
+                    new CodeInstruction(OpCodes.Ldarg_0),
+                    new CodeInstruction(OpCodes.Ldfld, Field(typeof(MapToolController), nameof(MapToolController.photonView))),
+                    isLocalOrVRPlayerInstr
+                )
+                // Only reset transforms conditionally (when local player is not in VR)
                 .MatchForward(false, new CodeMatch(OpCodes.Call, PropertyGetter(typeof(Vector3), nameof(Vector3.zero))))
-                .Advance(-3)
-                .RemoveInstructions(16)
+                .Advance(-2)
+                .SetAndAdvance(OpCodes.Call, ((Action<MapToolController>)ConditionalResetTransform).Method)
+                .RemoveInstructions(14)
                 .InstructionEnumeration()
             : new CodeMatcher(instructions)
                 .MatchForward(false, new CodeMatch(OpCodes.Ldc_R4, 90f))
@@ -178,19 +201,35 @@ internal static class UniversalMapToolPatches
                 .Insert(new CodeInstruction(OpCodes.Call, ((Func<MapToolController, float>)GetHoldAngle).Method))
                 .MatchForward(false,
                     new CodeMatch(OpCodes.Callvirt, PropertyGetter(typeof(PhotonView), nameof(PhotonView.IsMine))))
-                .SetInstructionAndAdvance(shouldMutate)
+                .SetInstructionAndAdvance(shouldMutateInstr)
                 .SetOpcodeAndAdvance(OpCodes.Brfalse_S)
                 .MatchForward(false,
                     new CodeMatch(OpCodes.Callvirt, PropertyGetter(typeof(PhotonView), nameof(PhotonView.IsMine))))
-                .SetInstructionAndAdvance(shouldMutate)
+                .SetInstructionAndAdvance(shouldMutateInstr)
                 .SetOpcodeAndAdvance(OpCodes.Brfalse_S)
                 .MatchForward(false,
                     new CodeMatch(OpCodes.Ldfld,
                         Field(typeof(MapToolController), nameof(MapToolController.FollowTransformClient))))
                 .Advance(-6) // PhotonView::get_IsMine
-                .SetInstructionAndAdvance(shouldMutate)
+                .SetInstructionAndAdvance(shouldMutateInstr)
                 .SetOpcodeAndAdvance(OpCodes.Brfalse_S)
                 .InstructionEnumeration();
+
+        static bool IsLocalOrVRPlayer(bool isLocal, PhotonView view)
+        {
+            return isLocal || NetworkSystem.instance.IsVRView(view);
+        }
+
+        // Only reset transforms if current player is not in VR
+        static void ConditionalResetTransform(MapToolController controller)
+        {
+            if (VRSession.InVR)
+                return;
+
+            controller.transform.parent.localPosition = Vector3.zero;
+            controller.transform.parent.localRotation = Quaternion.identity;
+            controller.mainSpringTransform.localRotation = Quaternion.identity;
+        }
 
         static bool ShouldMutateTransforms(PhotonView view)
         {