+++ Source/Utils.cs
@@ -1,10 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Security.Cryptography;
 using System.Text;
 using System.Text.RegularExpressions;
+using HarmonyLib;
 using RepoXR.Managers;
+using RepoXR.Networking;
 using Steamworks;
 using UnityEngine;
 
@@ -194,5 +197,29 @@ internal static class Utils
         private readonly float timeStarted = Time.realtimeSinceStartup;
 
         public override bool keepWaiting => !predicate() && Time.realtimeSinceStartup - timeStarted < timeout;
+    }
+}
+
+public static class PlayerLocalCameraExtensions
+{
+    public static MethodInfo GetHandOverrideTransformMethod =>
+        AccessTools.Method(typeof(PlayerLocalCameraExtensions), nameof(GetHandOverrideTransform));
+
+    public static Transform GetHandOverrideTransform(this PlayerLocalCamera camera)
+    {
+        // If overriden, always return the override transform
+        if (camera.GetOverrideActive())
+            return camera.playerAvatar.PlayerVisionTarget.VisionTransform;
+
+        // If we are in VR and the camera is local, return our VR hand transform
+        if (VRSession.Instance is { } instance && camera.playerAvatar.isLocal)
+            return instance.Player.MainHand;
+
+        // If the player is a VR player, return their VR hand transform
+        if (NetworkSystem.instance && NetworkSystem.instance.GetNetworkPlayer(camera.playerAvatar, out var player))
+            return player.PrimaryHand;
+
+        // Fallback, return the camera transform
+        return camera.transform;
     }
 }